Gidanbanta/backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Redis & Caching
redis==5.2.0
hiredis==3.0.0

# Real-time (Socket.IO)
python-socketio==5.11.4
python-engineio==4.10.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.2.1

# Background Tasks
celery==5.4.0
redis==5.2.0

# HTTP Client (for payment APIs and Football API)
httpx==0.27.2
requests==2.32.3

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Rate Limiting
slowapi==0.1.9

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
